pyproject.toml
[project]
name = "mcpdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.3",
    "langchain-groq>=1.1.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "mcp>=1.25.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
]
